Move function to usage
